+++ src/main/scala-2.11/ai/tripl/arc/jupyter/ArcInterpreter.scala
@@ -56,6 +56,7 @@ final class ArcInterpreter extends Interpreter {
   var confStreamingDuration = 10
   var confStreamingFrequency = 1000
   var confMonospace = false
+  var confLeftAlign = false
   var udfsRegistered = false
 
   var isJupyterLab: Option[Boolean] = None
@@ -136,7 +137,7 @@ final class ArcInterpreter extends Interpreter {
         // detect jupyterlab
         val jupyterLab = isJupyterLab.getOrElse(
           scala.util.Properties.envOrNone("JUPYTER_ENABLE_LAB") match {
-            case Some(j) => if (j == "yes") true else false
+            case Some(j) if (j == "yes") => true
             case None => false
           }
         )
@@ -170,7 +171,7 @@ final class ArcInterpreter extends Interpreter {
                 }.mkString(",")
               }
               case None => ""
-            }      
+            }
 
             ("arc", parseArgs(lines(0)),
               if (lines(0).startsWith("%sqlvalidate")) {
@@ -181,7 +182,7 @@ final class ArcInterpreter extends Interpreter {
                 |  "environments": [],
                 |  "sql": \"\"\"${lines.drop(1).mkString("\n")}\"\"\",
                 |  "sqlParams": {${sqlParams}},
-                |  ${commandArgs.filterKeys{ !List("name", "description", "sqlParams", "environments", "numRows", "truncate", "persist", "monospace", "streamingDuration").contains(_) }.map{ case (k, v) => s""""${k}": "${v}""""}.mkString(",")}
+                |  ${commandArgs.filterKeys{ !List("name", "description", "sqlParams", "environments", "numRows", "truncate", "persist", "monospace", "leftAlign", "streamingDuration").contains(_) }.map{ case (k, v) => s""""${k}": "${v}""""}.mkString(",")}
                 |}""".stripMargin
               } else {
                 s"""{
@@ -193,7 +194,7 @@ final class ArcInterpreter extends Interpreter {
                 |  "outputView": "${commandArgs.getOrElse("outputView", randStr(32))}",
                 |  "persist": ${commandArgs.getOrElse("persist", "false")},
                 |  "sqlParams": {${sqlParams}}
-                |  ${commandArgs.filterKeys{ !List("name", "description", "sqlParams", "environments", "outputView", "numRows", "truncate", "persist", "monospace", "streamingDuration").contains(_) }.map{ case (k, v) => s""""${k}": "${v}""""}.mkString(",")}
+                |  ${commandArgs.filterKeys{ !List("name", "description", "sqlParams", "environments", "outputView", "numRows", "truncate", "persist", "monospace", "leftAlign", "streamingDuration").contains(_) }.map{ case (k, v) => s""""${k}": "${v}""""}.mkString(",")}
                 |}""".stripMargin
               }
             )
@@ -245,6 +246,7 @@ final class ArcInterpreter extends Interpreter {
         val streamingDuration = Try(commandArgs.get("streamingDuration").get.toInt).getOrElse(confStreamingDuration)
         val persist = Try(commandArgs.get("persist").get.toBoolean).getOrElse(false)
         val monospace = Try(commandArgs.get("monospace").get.toBoolean).getOrElse(confMonospace)
+        val leftAlign = Try(commandArgs.get("leftAlign").get.toBoolean).getOrElse(confLeftAlign)
 
         // store previous values so that the ServiceLoader resolution is not called each run
         val pipelineStagePlugins = memoizedPipelineStagePlugins match {
@@ -297,7 +299,7 @@ final class ArcInterpreter extends Interpreter {
         outputHandler match {
           case Some(outputHandler) => {
             interpreter match {
-              case "arc" | "summary" => {
+              case "arc" | "summary" | "cypher" => {
                 val listener = new ProgressSparkListener(listenerElementHandle, jupyterLab)(outputHandler, logger)
                 listener.init()(outputHandler)
                 spark.sparkContext.addSparkListener(listener)
@@ -327,7 +329,7 @@ final class ArcInterpreter extends Interpreter {
                       case _ => {
                         ARC.run(pipeline) match {
                           case Some(df) => {
-                            val result = renderResult(outputHandler, df, numRows, truncate, streamingDuration)
+                            val result = renderResult(outputHandler, df, numRows, truncate, monospace, leftAlign, streamingDuration)
                             memoizedUserData = arcContext.userData
                             result
                           }
@@ -349,7 +351,7 @@ final class ArcInterpreter extends Interpreter {
             val config = ConfigFactory.parseString(s"""{"plugins": {"config": [${command}]}}""", ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF))
             val dynamicConfigsOrErrors = ai.tripl.arc.config.Plugins.resolveConfigPlugins(config, "plugins.config", arcContext.dynamicConfigurationPlugins)(spark, logger, arcContext)
             dynamicConfigsOrErrors match {
-              case Left(errors) => ExecuteResult.Error(ai.tripl.arc.config.Error.pipelineSimpleErrorMsg(errors, false))
+                case Left(errors) => ExecuteResult.Error(ai.tripl.arc.config.Error.pipelineSimpleErrorMsg(errors, false))
               case Right(dynamicConfigs) => {
                 val dynamicConfigsConf = dynamicConfigs.reduceRight[Config]{ case (c1, c2) => c1.withFallback(c2) }
                 val entryMap = dynamicConfigsConf.entrySet.asScala.map { entry =>
@@ -378,7 +380,7 @@ final class ArcInterpreter extends Interpreter {
             }
             if (persist) df.persist(StorageLevel.MEMORY_AND_DISK_SER)
             ExecuteResult.Success(
-              DisplayData.html(renderHTML(df, numRows, truncate))
+              DisplayData.html(renderHTML(df, numRows, truncate, monospace, leftAlign))
             )
           }
           case "printmetadata" => {
@@ -394,7 +396,7 @@ final class ArcInterpreter extends Interpreter {
             }
             if (persist) df.persist(StorageLevel.MEMORY_AND_DISK_SER)
             ExecuteResult.Success(
-              DisplayData.html(renderHTML(df, numRows, truncate))
+              DisplayData.html(renderHTML(df, numRows, truncate, monospace, leftAlign))
             )
           }
           case "env" => {
@@ -440,14 +442,23 @@ final class ArcInterpreter extends Interpreter {
             commandArgs.get("monospace") match {
               case Some(monospace) => {
                 try {
-                  val monospaceValue = monospace.toBoolean
-                  confMonospace = monospaceValue
+                  confMonospace = monospace.toBoolean
                 } catch {
                   case e: Exception =>
                 }
               }
               case None =>
-            }             
+            }
+            commandArgs.get("leftAlign") match {
+              case Some(leftAlign) => {
+                try {
+                  confLeftAlign = leftAlign.toBoolean
+                } catch {
+                  case e: Exception =>
+                }
+              }
+              case None =>
+            }
             commandArgs.get("streamingDuration") match {
               case Some(streamingDuration) => {
                 try {
@@ -460,13 +471,17 @@ final class ArcInterpreter extends Interpreter {
               case None =>
             }
             val text = s"""
+            |Arc Options:
             |master: ${confMaster}
             |memory: ${runtimeMemorySize}B
-            |monospace: ${confMonospace}
-            |numRows: ${confNumRows}
-            |truncate: ${confTruncate}
             |streaming: ${confStreaming}
             |streamingDuration: ${confStreamingDuration}
+            |
+            |Display Options:
+            |numRows: ${confNumRows}
+            |truncate: ${confTruncate}
+            |leftAlign: ${leftAlign}
+            |monospace: ${confMonospace}
             """.stripMargin
             ExecuteResult.Success(
               DisplayData.text(text)
@@ -510,10 +525,10 @@ final class ArcInterpreter extends Interpreter {
     }
   }
 
-  def renderResult(outputHandler: Option[OutputHandler], df: DataFrame, numRows: Int, truncate: Int, streamingDuration: Int) = {
+  def renderResult(outputHandler: Option[OutputHandler], df: DataFrame, numRows: Int, truncate: Int, monospace: Boolean, leftAlign: Boolean, streamingDuration: Int) = {
     if (!df.isStreaming) {
       ExecuteResult.Success(
-        DisplayData.html(renderHTML(df, numRows, truncate))
+        DisplayData.html(renderHTML(df, numRows, truncate, monospace, leftAlign))
       )
     } else {
       outputHandler match {
@@ -545,13 +560,13 @@ final class ArcInterpreter extends Interpreter {
               // create the html handle on the first run
               if (initial) {
                 outputHandler.html(
-                  renderHTML(df, numRows, truncate),
+                  renderHTML(df, numRows, truncate, monospace, leftAlign),
                   outputElementHandle
                 )
                 initial = false
               } else {
                 outputHandler.updateHtml(
-                  renderHTML(df, numRows, truncate),
+                  renderHTML(df, numRows, truncate, monospace, leftAlign),
                   outputElementHandle
                 )
               }
@@ -569,7 +584,7 @@ final class ArcInterpreter extends Interpreter {
           writeStream.stop
           outputHandler.html("", outputElementHandle)
           ExecuteResult.Success(
-            DisplayData.html(renderHTML(spark.table(queryName), numRows, truncate))
+            DisplayData.html(renderHTML(spark.table(queryName), numRows, truncate, monospace, leftAlign))
           )
         }
         case None => ExecuteResult.Error("No result.")
@@ -580,7 +595,7 @@ final class ArcInterpreter extends Interpreter {
   def currentLine(): Int =
     count
 
-  def renderHTML(df: DataFrame, numRows: Int, truncate: Int): String = {
+  def renderHTML(df: DataFrame, numRows: Int, truncate: Int, monospace: Boolean, leftAlign: Boolean): String = {
     import xml.Utility.escape
 
     val header = df.columns
@@ -625,7 +640,10 @@ final class ArcInterpreter extends Interpreter {
       }: Seq[String]
     }
 
-    s"""<table><tr>${header.map(h => s"<th>${escape(h)}</th>").mkString}</tr>${rows.map { row => s"<tr>${row.map { cell => s"<td>${escape(cell)}</td>" }.mkString}</tr>"}.mkString}</table>"""
+    val monospaceClass = if (monospace) "monospace" else ""
+    val leftAlignClass = if (leftAlign) "leftalign" else ""
+
+    s"""<table class="${monospaceClass} ${leftAlignClass}"><tr>${header.map(h => s"<th>${escape(h)}</th>").mkString}</tr>${rows.map { row => s"<tr>${row.map { cell => s"<td>${escape(cell)}</td>" }.mkString}</tr>"}.mkString}</table>"""
   }
 
   def parseArgs(input: String): collection.mutable.Map[String, String] = {